requirements.txt
# Core
discord.py[voice]>=2.5.0
davey>=0.1.3
discord-ext-voice-recv>=0.5.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
websockets>=12.0

# Speech-to-Text
faster-whisper>=1.0.0

# Wake Word Detection
openwakeword>=0.6.0

# Text-to-Speech (local)
piper-tts>=1.2.0
pathvalidate>=3.0.0  # required by piper-tts but not declared as dependency

# Audio Processing
numpy>=1.24.0
pydub>=0.25.1
scipy>=1.11.0

# Optional: ElevenLabs TTS
# elevenlabs>=1.0.0
